refactor: AlarmSection 컴포넌트 분리
